Add block allocation function
Adds a function that takes block(s) of memory and constructs the tree using that, as well as a wrapper function that does the block allocation and calls the constructor function. Appropriate tests are included. Also includes a refactor of expect functions from the testing file into a util file.